fix(v2client): fix v2 client org header
